% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{associate}
\usage{
associate(x, y, yLab = NULL)
}
\arguments{
\item{x}{Data frame containing the explanatory variable data}

\item{y}{Data frame containing the response data}

\item{yLab}{Character string containing label for the response variable}
}
\description{
\code{associate} Performs the associate analysis between a set of
independent variables (x) and a dependent variable (y).  Conducts several
ANOVA tests of independence
}
\seealso{
Other regression functions: \code{\link{conclusion}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{prediction}},
  \code{\link{regressionAnalysis}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
