% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestPredictions.R
\name{bestPredictions}
\alias{bestPredictions}
\title{bestPredictions}
\usage{
bestPredictions(best, models, newdata)
}
\arguments{
\item{best}{Data frame containing the top performing models and estimators}

\item{models}{List of models}

\item{newdata}{Data frame containing new data to predict.}
}
\value{
predictions Data frame containing predictions for newdata
}
\description{
\code{bestPredictions} Performs predictions using select models on new data.
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaComplexity}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@yXsciencesalon.org}
}
