% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{bivariate}
\alias{bivariate}
\title{bivariate}
\usage{
bivariate(data)
}
\arguments{
\item{data}{Data frame containing data to be analyzed}
}
\description{
\code{bivariate} Performs the bivariate analysis of independent variables on audience score.
}
\seealso{
Other EDA functions: \code{\link{bivariateQual}},
  \code{\link{bivariateQuant}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
