% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateQual.R
\name{bivariateQual}
\alias{bivariateQual}
\title{bivariateQual}
\usage{
bivariateQual(data, xLab)
}
\arguments{
\item{data}{Two column data frame, containing the dependent and independent variables.}

\item{xLab}{Capitalized character string for the variable name or label}
}
\value{
analysis List containing:
\itemize{
 \item Data frame containing summary statistics by categorical level.
 \item Boxplot showing the distribution of dependent variable by categorical level.
 \item Anova test of equal means.
}
}
\description{
\code{bivariateQual} Performs a bivariate analysis of categorical independent
variables vis-a-vis the quantitative dependent variable.  Analyses include
the presentation of box plots, summary statistics by categorical level, and
ANOVA tests of the difference in mean values of the dependent variable by
categorical level.
}
\seealso{
Other EDA functions: \code{\link{bivariateQuant}},
  \code{\link{bivariate}}, \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
