% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateQuant.R
\name{bivariateQuant}
\alias{bivariateQuant}
\title{bivariateQuant}
\usage{
bivariateQuant(data, xLab)
}
\arguments{
\item{data}{Two column data frame, containing the dependent and independent variables.}

\item{xLab}{Capitalized character string for the variable name or label}
}
\value{
analysis List containing:
\itemize{
 \item Scatterplot evincing the relationship between the independent and dependent variables.
 \item Correlation test.
}
}
\description{
\code{bivariateQuant} Performs a bivariate analysis of quantitative independent
variables vis-a-vis the quantitative dependent variable.  Analyses include
the presentation of scatter plots, and correlation tests.
}
\seealso{
Other EDA functions: \code{\link{bivariateQual}},
  \code{\link{bivariate}}, \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
