% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{bma}
\usage{
bma(yX)
}
\arguments{
\item{yX}{Data frame containing the vector y and matrix X of parameters.}
}
\value{
list of BMA models
}
\description{
\code{bma} Performs Bayes Model Averaging using several default priors
}
\seealso{
Other BMA functions: \code{\link{bmaCompare}},
  \code{\link{bmaPredict}}, \code{\link{bmaReport}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
