% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaBest.R
\name{bmaBest}
\alias{bmaBest}
\title{bmaBest}
\usage{
bmaBest(yX)
}
\arguments{
\item{yX}{Data frame containing the vector y and matrix X of parameters.}
}
\value{
list of BMA models
}
\description{
\code{bmaBest} Performs Bayes Model Averaging for select priors
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
