% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaCompare.R
\name{bmaCompare}
\alias{bmaCompare}
\title{bmaCompare}
\usage{
bmaCompare(yX, iterations = 2)
}
\arguments{
\item{yX}{Data frame containing the y vector and X matrix of parameters.}

\item{iterations}{Numeric indicating the number of iterations to perform}
}
\description{
\code{bmaCompare} Compares predictive accuracy of best predictive models
}
\seealso{
Other BMA functions: \code{\link{bmaPredict}},
  \code{\link{bmaReport}}, \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
