% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaComplexity.R
\name{bmaComplexity}
\alias{bmaComplexity}
\title{bmaComplexity}
\usage{
bmaComplexity(models)
}
\arguments{
\item{models}{List of BMA models}
}
\value{
list of reports
}
\description{
\code{bmaComplexity} Plots model complexity vis-a-vis posterior probability
}
\seealso{
Other BMA functions: \code{\link{bestPredictions}},
  \code{\link{bmaAnalysis}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
