% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaEvaluation.R
\name{bmaEvaluation}
\alias{bmaEvaluation}
\title{bmaEvaluation}
\usage{
bmaEvaluation(mList, candidates)
}
\arguments{
\item{mList}{List of candidate BAS.lm models}

\item{candidates}{Dataframe containing the candidate models to be evaluated.}
}
\description{
\code{bmaEvaluation} Prepares data by which the designated models would be evaluated.
}
\seealso{
Other BMA functions: \code{\link{bestPredictions}},
  \code{\link{bmaAnalysis}}, \code{\link{bmaComplexity}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
