% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaModel1.R
\name{bmaModel1}
\alias{bmaModel1}
\title{bmaModel1}
\usage{
bmaModel1(models)
}
\arguments{
\item{models}{List of BMA models}
}
\value{
A list containing:
\itemize{
 \item m1s Data frame containing posterior model probabilities, R2 scores and model sizes for each prior.
 \item m1Plots - List of plots summarizing posterior probabilities, R2, and model sizes across each prior. 
}
}
\description{
\code{bmaModel1} Summarizes model 1 under BMA for a series of models
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaPDC}},
  \code{\link{bmaPIPPlots}}, \code{\link{bmaPIP}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
