% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaModel1Plots.R
\name{bmaModel1Plots}
\alias{bmaModel1Plots}
\title{bmaModel1Plots}
\usage{
bmaModel1Plots(m1s)
}
\arguments{
\item{m1s}{Data frame containing the following Model 1 summary information:
\itemize{
 \item Prior - Character string containing the name of the prior
 \item Posterior Probability Numeric - containing the posterior probability of the top model for each prior
 \item R2 - Numeric representing the R2 value for the top model for each prior.
 \item Size - Integer indicating the number of parameters in the top model
}}
}
\value{
list of containing the following plots:
\itemize{
 \item postProbs - Plot of posterior model probabilities for the top model of each prior.
 \item r2 - Plot of the R2 score for the top model of each prior.
 \item size - Plot containing the model size for the top model of each prior. 
}
}
\description{
\code{bmaModel1Plots} Renders plots for BMA model 1 for a series of models
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaBest}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1}}, \code{\link{bmaPDC}},
  \code{\link{bmaPIPPlots}}, \code{\link{bmaPIP}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
