% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPDC.R
\name{bmaPDC}
\alias{bmaPDC}
\title{bmaPDC}
\usage{
bmaPDC(m, estimator = NULL)
}
\arguments{
\item{m}{A BAS.lm object.}

\item{estimator}{Estimator to be used when calculating the coefficients}
}
\value{
List with descriptive statistics and credible intervals for the means
}
\description{
\code{bmaPDC} Reports the posterior distribution of coefficients for a model
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPIPPlots}}, \code{\link{bmaPIP}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
