% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPerformance.R
\name{bmaPerformance}
\alias{bmaPerformance}
\title{bmaPerformance}
\usage{
bmaPerformance(yX, trials = 3)
}
\arguments{
\item{yX}{Data frame containing the y vector and X matrix of parameters.}

\item{trials}{Numeric indicating the number of trials to perform}
}
\description{
\code{bmaPerformance} Compares predictive accuracy of best predictive models
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaBest}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
