% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPerformanceReport.R
\name{bmaPerformanceReport}
\alias{bmaPerformanceReport}
\title{bmaPerformanceReport}
\usage{
bmaPerformanceReport(performance)
}
\arguments{
\item{performance}{List containing performance data from \code{bmaPerformance}}
}
\description{
\code{bmaPerformanceReport} Renders tables and plots for BMA Compare function
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaBest}}, \code{\link{bmaComplexity}},
  \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
