% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPIPPlots.R
\name{bmaPIPPlots}
\alias{bmaPIPPlots}
\title{bmaPIPPlots}
\usage{
bmaPIPPlots(pip)
}
\arguments{
\item{pip}{Data frame containing parameter inclusion probabilities from \code{bmaPIP.R}.}
}
\value{
list of plots of parameter inclusion probabilities across models.
}
\description{
\code{bmaPIPPlots} Renders tables and plots of parameter inclusing probabilities
}
\details{
Renders tables and plots of parameter inclusion probabilities by model variable. 
The table summarizes parameter inclusion probabilities across priors, highlighting
the probabilities including 0.5. Each plot reports inclusion probabilities 
of a single predictor across the  series of models.
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIP}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
