% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPredictModels.R
\name{bmaPredictModels}
\alias{bmaPredictModels}
\title{bmaPredictModels}
\usage{
bmaPredictModels(models, yX, trial = NULL)
}
\arguments{
\item{models}{List of BMA models}

\item{yX}{Data frame containing test data}

\item{trial}{Numeric indicator of the number of trials of repeated predictions}
}
\value{
MSE frame containing MSE measures for each model
}
\description{
\code{bmaPredictModels} Predictions cross validation
}
\seealso{
Other BMA functions: \code{\link{bestPredictions}},
  \code{\link{bmaAnalysis}}, \code{\link{bmaComplexity}},
  \code{\link{bmaEvaluation}}, \code{\link{bmaPDC}},
  \code{\link{bmaPIPPlots}}, \code{\link{bmaPIP}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@yXsciencesalon.org}
}
