% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaReport.R
\name{bmaReport}
\alias{bmaReport}
\title{bmaReport}
\usage{
bmaReport(models, top = TRUE)
}
\arguments{
\item{models}{List of BMA models}
}
\value{
list of reports
}
\description{
\code{bmaReport} Reports results from a series of BMA Regression for a series of models.
}
\seealso{
Other BMA functions: \code{\link{bmaCompare}},
  \code{\link{bmaPredict}}, \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
