% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAssociation.R
\name{cramers}
\alias{cramers}
\title{cramers}
\usage{
cramers(data)
}
\arguments{
\item{data}{Data frame or vector containing a single quantitative variable}
}
\value{
Data frame containing the Cramer's V values for the associations between 
each of the categorical variables in the data set.
}
\description{
\code{cramers} Cramer's V test of strength of association
}
\seealso{
Other analysis functions: \code{\link{getCSampleSize}},
  \code{\link{getPSampleSize}},
  \code{\link{getSummaryStats}}, \code{\link{pearsons}},
  \code{\link{x2}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
