% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{forward}
\usage{
forward(data, y)
}
\arguments{
\item{data}{Data frame containing the full model}

\item{y}{Character string containing the name of the response variable}
}
\description{
\code{forward} Creates prediction model using forward selection
}
\seealso{
Other regression functions: \code{\link{associate}},
  \code{\link{conclusion}}, \code{\link{correlate}},
  \code{\link{evalForwardModel}}, \code{\link{prediction}},
  \code{\link{regressionAnalysis}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
