% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getCSampleSize}
\alias{getCSampleSize}
\title{getCSampleSize}
\usage{
getCSampleSize(data, e, p = 0.8, me = 0.05)
}
\arguments{
\item{e}{Numeric indicator of effect size}

\item{p}{Numeric indicator of the target power}

\item{me}{Numeric indicator of alpha level}
}
\value{
sampleSize Minimum required sample size
}
\description{
\code{getCSampleSize} Computes the required sample size for a proportion, such
that the population proportion is within a designated margin of error of the
sample proportion with a designated level of confidence.
}
\seealso{
Other analysis functions: \code{\link{cramers}},
  \code{\link{getPSampleSize}},
  \code{\link{getSummaryStats}}, \code{\link{pearsons}},
  \code{\link{x2}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
