% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getPSampleSize}
\alias{getPSampleSize}
\title{getPSampleSize}
\usage{
getPSampleSize(p, conf = 0.95, me = 0.05)
}
\arguments{
\item{p}{Numeric proportion between 0 and 1}

\item{conf}{Numeric confidence level between 0 and 1}

\item{me}{Numeric margin of error between 0 and 1}
}
\value{
sampleSize Minimum required sample size
}
\description{
\code{getPSampleSize} Computes the required sample size for a proportion, such
that the population proportion is within a designated margin of error of the
sample proportion with a designated level of confidence.
}
\seealso{
Other analysis functions: \code{\link{cramers}},
  \code{\link{getCSampleSize}},
  \code{\link{getSummaryStats}}, \code{\link{pearsons}},
  \code{\link{x2}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
