% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getSummaryStats}
\alias{getSummaryStats}
\title{getSummaryStats}
\usage{
getSummaryStats(data, xLab = NULL)
}
\arguments{
\item{data}{Data frame or vector containing a single quantitative variable}

\item{xLab}{Character string containing the name of the variable or categorical level}
}
\value{
Data frame containing summary statistics
}
\description{
\code{getSummaryStats} Provides summary stats for quantitative variable
}
\seealso{
Other analysis functions: \code{\link{getCSampleSize}},
  \code{\link{getPSampleSize}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
