% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\name{intro}
\alias{intro}
\title{intro}
\usage{
intro(data)
}
\arguments{
\item{data}{Data frame containing the full model#'}
}
\value{
Data frame and histogram
}
\description{
\code{intro} Provides descriptive statistics for intro section.
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
