% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAssociation.R
\name{pearsons}
\alias{pearsons}
\title{pearsons}
\usage{
pearsons(data)
}
\arguments{
\item{data}{Data frame or vector containing a single quantitative variable}
}
\value{
Correlation plot
}
\description{
\code{pearsons} Pearsons coefficients are computed to measure correlation
among quantitative variables
}
\seealso{
Other analysis functions: \code{\link{cramers}},
  \code{\link{getCSampleSize}},
  \code{\link{getPSampleSize}},
  \code{\link{getSummaryStats}}, \code{\link{x2}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
