% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBar2}
\alias{plotBar2}
\title{plotBar2}
\usage{
plotBar2(data, yLab, xLab, plotTitle = NULL, values = FALSE)
}
\arguments{
\item{data}{Data frame or vector containing a single categorical factor variable}

\item{yLab}{Capital case character string describing the y variable}

\item{xLab}{Capital case character string containing the name of the variable x variable}

\item{plotTitle}{Capital case character string for the title of the plot}
}
\value{
Bar plot
}
\description{
\code{plotBar2} Renders a bar plot
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCorr}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResFit}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
