% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCIBars}
\alias{plotCIBars}
\title{plotCIBars}
\usage{
plotCIBars(data, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame containing terms, means, and lower and upper bounds of confidence interval}

\item{plotTitle}{Capital case character string for the title of the plot}
}
\value{
confIntPlot
}
\description{
\code{plotCIBars} Plots confidence interval bars
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLine}}, \code{\link{plotQQ}},
  \code{\link{plotResFit}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
