% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCooks}
\alias{plotCooks}
\title{plotCooks}
\usage{
plotCooks(mod, mName)
}
\arguments{
\item{mod}{Linear model object}

\item{mName}{Capital case character string for the model name}
}
\value{
cooks Cooks distance plot for residuals
}
\description{
\code{plotCooks} Plots cooks distance
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCorr}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLinear}},
  \code{\link{plotLine}}, \code{\link{plotQQ}},
  \code{\link{plotResAll}}, \code{\link{plotResFit}},
  \code{\link{plotResLeverage}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
