% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCorr}
\alias{plotCorr}
\title{plotCorr}
\usage{
plotCorr(data, yVar)
}
\arguments{
\item{data}{Data frame containing data for which the correlations will be plotted.}
}
\value{
Correlation plot
}
\description{
\code{plotCorr} Creates correlation plot
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar2}},
  \code{\link{plotBar}}, \code{\link{plotBox}},
  \code{\link{plotCIBars}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResFit}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
