% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFreqProp}
\alias{plotFreqProp}
\title{plotFreqProp}
\usage{
plotFreqProp(data, yLab = "Movies", xLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame containing the variables:
\itemize{
 \item Character string containing the values for each categorical level
 \item N: The number of observations for the categorical level
 \item Proportion: The proportion of observation for the categorical level
 \item Cumulative: The cumulative proportion
 \item Pos: The y position for the frequency and proportion text
}}

\item{yLab}{Capital case character string describing the y variable}

\item{xLab}{Capital case character string containing the name of the variable x variable}

\item{plotTitle}{Capital case character string for the title of the plot}
}
\value{
List containing a contingency table and a stacked bar plot.
}
\description{
\code{plotFreqPropTbl} Renders a stacked bar plot and a contingency table
showing frequencies and proportions
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar2}},
  \code{\link{plotBar}}, \code{\link{plotBox}},
  \code{\link{plotCIBars}}, \code{\link{plotCorr}},
  \code{\link{plotHist}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResFit}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
