% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotHist}
\alias{plotHist}
\title{plotHist}
\usage{
plotHist(data, xLab = NULL, yLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame containing a single numeric variable}

\item{xLab}{Capital case character string the group or subset of data being printed (optional)}

\item{yLab}{Capital case character string describing y is being printed}

\item{plotTitle}{Capital case character string for the plotTitle of the plot}
}
\value{
List containing a summary statistics and a histogram
}
\description{
\code{plotHist} Renders a histogram with a normal curve as well as a table
of summary statistics.
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar2}},
  \code{\link{plotBar}}, \code{\link{plotBox}},
  \code{\link{plotCIBars}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResFit}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}},
  \code{\link{plotStackedBar}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
