% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLinear}
\alias{plotLinear}
\title{plotLinear}
\usage{
plotLinear(mod, yVar, yLab = NULL)
}
\arguments{
\item{mod}{Linear model object}

\item{yVar}{Character string for the dependent variable}

\item{yLab}{Character string for the y-label}
}
\value{
list containing ressidual v predictor plots
}
\description{
\code{plotLinear} Plots linear regression lines for each predictor
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCooks}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLine}}, \code{\link{plotQQ}},
  \code{\link{plotResAll}}, \code{\link{plotResFit}},
  \code{\link{plotResLeverage}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
