% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotResFit}
\alias{plotResFit}
\title{plotResFit}
\usage{
plotResFit(mod, yLab)
}
\arguments{
\item{mod}{Linear model object}

\item{yLab}{Capital case character string for response variable}
}
\value{
resFit Residual vs. Fit Plot
}
\description{
\code{plotResFit} Plots regression diagnostics
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCorr}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
