% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotResLeverage}
\alias{plotResLeverage}
\title{plotResLeverage}
\usage{
plotResLeverage(mod, mName)
}
\arguments{
\item{mod}{Linear model object}

\item{mName}{Capital case character string for the model name}
}
\value{
resLeverage Residual vs. Leverage Plot
}
\description{
\code{plotResLeverage} Plots regression diagnostics
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCooks}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLinear}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
