% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotResQQ}
\alias{plotResQQ}
\title{plotResQQ}
\usage{
plotResQQ(res, mName)
}
\arguments{
\item{res}{Vector of residual values}

\item{mName}{Capital case character string for the model name}
}
\value{
resQQ Residual Normal QQ Plot
}
\description{
\code{plotResQQ} Plots regression diagnostics
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar2}},
  \code{\link{plotBar}}, \code{\link{plotBox}},
  \code{\link{plotCIBars}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLine}}, \code{\link{plotQQ}},
  \code{\link{plotResFit}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
