% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotScatter}
\alias{plotScatter}
\title{plotScatter}
\usage{
plotScatter(data, xLab, yLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame containing the quantitative variables}

\item{xLab}{Capital case character string containing the name of the grouping or subset variable}

\item{yLab}{Capital case character string containing the name of the y variable}

\item{plotTitle}{Capital case character string for title of plot}
}
\value{
Scatterplot object
}
\description{
\code{plotScatter} Renders a scatterplot for two numerical variablesa
}
\seealso{
Other visualization functions: \code{\link{groupBarPlot}},
  \code{\link{pieChart}}, \code{\link{plotBar2}},
  \code{\link{plotBar}}, \code{\link{plotBox}},
  \code{\link{plotCIBars}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLine}}, \code{\link{plotQQ}},
  \code{\link{plotResFit}}, \code{\link{plotResQQ}},
  \code{\link{plotStackedBar}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
