% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(mod, cases, conf = 0.95)
}
\arguments{
\item{mod}{Linear regression model}

\item{cases}{Data frame containing cases to predict}

\item{conf}{Numeric indicator of desired confidence}

\item{mName}{Character string containing the model name}
}
\value{
Data frame with summary of predictions
}
\description{
\code{prediction} Performs prediction on an randomly sampled individual observation
}
\seealso{
Other regression functions: \code{\link{associate}},
  \code{\link{conclusion}}, \code{\link{correlate}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{regressionAnalysis}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
