% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionAnalysis.R
\name{regressionAnalysis}
\alias{regressionAnalysis}
\title{regressionAnalysis}
\usage{
regressionAnalysis(mod, mName, yVar, yLab, full = TRUE)
}
\arguments{
\item{mod}{Linear model}

\item{mName}{Capitalized character string for the name of the linear model}

\item{yVar}{Character string containing the name of the dependent variable}

\item{yLab}{Capitalized character string for the dependent variable}

\item{full}{Logical indicating whether a full analysis should be conducted}
}
\value{
analysis plots, summary statistics, test results and interpretive text
}
\description{
\code{regressionAnalysis} Performs regression analysis
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
