% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr}
\alias{slr}
\title{slr}
\usage{
slr(data, y)
}
\arguments{
\item{data}{Data frame containing movie data}

\item{y}{Character string containing the name of the response variable}
}
\value{
Data frame containing summary statistics for each model
}
\description{
\code{slr} Performs simple linear regression of all quantitative variables
on the designated response variable.
}
\seealso{
Other analysis functions: \code{\link{getCSampleSize}},
  \code{\link{getPSampleSize}},
  \code{\link{getSummaryStats}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
