% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(data)
}
\arguments{
\item{data}{Data frame containing data for analysis}
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other EDA functions: \code{\link{bivariateQual}},
  \code{\link{bivariateQuant}}, \code{\link{bivariate}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
