% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateQual.R
\name{univariateQual}
\alias{univariateQual}
\title{univariateQual}
\usage{
univariateQual(data, type = "dichotomous", groups, xLab, yLab = NULL,
  plotTitle = NULL)
}
\arguments{
\item{data}{Wide data frame containing:
\itemize{
 \item Character string containing the grouping variale
 \item Character string containing the category level
 \item Numeric containing the frequencies by category level
 \item Numeric indicating the proportions by category level
 \item Numeric containing the cumulative proportion by group and category level
 \item pos the vertical position of the labels on the bars
}}

\item{type}{Character string indicating whether data is dichotomous, ordinal, or nominal}

\item{groups}{Character vector containing the group names to be rendered
along the x-axis}

\item{xLab}{Character string containing the label for the x-axis}

\item{yLab}{Character string indicating the label for the y-axis}

\item{plotTitle}{Character string indicating the plot title.}
}
\value{
analysis List containing:
 1. Contingency table
 2. Stacked Frequency Proportion Barplot
}
\description{
\code{univariateQual} Performs univariate analysis on a group of dichotomous
qualitative variables with uniform category levels. The function returns
a combined contingency table for all groups and categories, as well as
a stacked bar plot showing frequencies and percentages of categories
for each group.
}
\seealso{
Other EDA functions: \code{\link{bivariateQual}},
  \code{\link{bivariateQuant}}, \code{\link{bivariate}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
