% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAssociation.R
\name{x2}
\alias{x2}
\title{x2}
\usage{
x2(data)
}
\arguments{
\item{data}{Data frame or vector containing a single quantitative variable}
}
\value{
Data frame containing the p-values for the associations between 
each of the categorical variables in the data set.
}
\description{
\code{x2} Chi square tests are conducted to test the significance
of association between a set of categorical variables.
}
\seealso{
Other analysis functions: \code{\link{cramers}},
  \code{\link{getCSampleSize}},
  \code{\link{getPSampleSize}},
  \code{\link{getSummaryStats}}, \code{\link{pearsons}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
