% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaAnalysis.R, R/postModelProb.R
\name{bmaAnalysis}
\alias{bmaAnalysis}
\alias{bmaAnalysis}
\title{bmaAnalysis}
\usage{
bmaAnalysis(models, top = FALSE)

bmaAnalysis(models, top = FALSE)
}
\arguments{
\item{models}{List of BMA models}

\item{models}{List of BMA models}
}
\value{
list of reports

list of reports
}
\description{
\code{bmaAnalysis} Reports results from a series of BMA Regression for a series of models.

\code{bmaAnalysis} Reports results for a series of models.
}
\seealso{
Other BMA functions: \code{\link{bmaBest}},
  \code{\link{bmaComplexity}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}

Other BMA functions: \code{\link{bmaBest}},
  \code{\link{bmaComplexity}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPIP}}, \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
