% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPIP.R
\name{bmaPIP}
\alias{bmaPIP}
\title{\code{bmaPIP} Summarizes parameter inclusing probabilities across models}
\usage{
bmaPIP(models)
}
\arguments{
\item{models}{List of BMA models}
}
\value{
Data frame containing parameter inclusion probabilities for each
model
}
\description{
\code{bmaPIP} Summarizes parameter inclusing probabilities across models
}
\seealso{
Other BMA functions: \code{\link{bmaAnalysis}},
  \code{\link{bmaBest}}, \code{\link{bmaEvaluation}},
  \code{\link{bmaModel1Plots}}, \code{\link{bmaModel1}},
  \code{\link{bmaPDC}}, \code{\link{bmaPIPPlots}},
  \code{\link{bmaPerformanceReport}},
  \code{\link{bmaPerformance}},
  \code{\link{bmaPredictModels}}, \code{\link{bmaPredict}},
  \code{\link{bma}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
