% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmaPredict.R
\name{bmaPredict}
\alias{bmaPredict}
\title{bmaPredict}
\usage{
bmaPredict(models, yX)
}
\arguments{
\item{models}{List of BMA models}

\item{yX}{Data frame containing test data}
}
\value{
mse frame containing MSE measures for each model
}
\description{
\code{bmaPredict} Predictions cross validation
}
\seealso{
Other BMA functions: \code{\link{bmaCompare}},
  \code{\link{bmaReport}}, \code{\link{bma}}
}
\author{
John James, \email{jjames@yXsciencesalon.org}
}
