% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLine}
\alias{plotLine}
\title{plotLine}
\usage{
plotLine(data, xLab = NULL, yLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame containing plokt data}

\item{xLab}{Capital case character string containing the name of the x variable (optional)}

\item{yLab}{Capital case character string containing the name of the y variable}

\item{plotTitle}{Character case character string containing the title for the plot}
}
\value{
List containing a contingency table and a stacked bar plot.
}
\description{
\code{plotLine} Renders a line plot
}
\seealso{
Other visualization functions: \code{\link{pieChart}},
  \code{\link{plotBar2}}, \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCIBars}},
  \code{\link{plotCorr}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotQQ}},
  \code{\link{plotResFit}}, \code{\link{plotResQQ}},
  \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
